Reorganize test targets.
