--- Functional Tests/Supporting Files/Fixtures/token.json
@@ -1,7 +1,0 @@
-{
-	"access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1NzkxMDE0NjMsImlhdCI6MTU3ODkyODY2MywiaW5zdGFuY2UiOiI5NzU1MTZmMS1mOWUzLTRlNTUtYTQ0ZC1lNDA3OTIzMmY5NDciLCJpc3MiOiJhcGlfa2V5cy80ZTQyOWNjNS0wM2YzLTQwNzctYmY4ZC04YTcxYWMwYWM2ODgiLCJyZWZyZXNoIjp0cnVlLCJzdWIiOiJib2IifQ.OTeVQ8KkX4ms98bsF6waJqcxAf7yMmCAVqFTX1GCOmc",
-	"refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1NzkxMDE0NjMsImlhdCI6MTU3ODkyODY2MywiaW5zdGFuY2UiOiI5NzU1MTZmMS1mOWUzLTRlNTUtYTQ0ZC1lNDA3OTIzMmY5NDciLCJpc3MiOiJhcGlfa2V5cy80ZTQyOWNjNS0wM2YzLTQwNzctYmY4ZC04YTcxYWMwYWM2ODgiLCJyZWZyZXNoIjp0cnVlLCJzdWIiOiJib2IifQ.OTeVQ8KkX4ms98bsF6waJqcxAf7yMmCAVqFTX1GCOmc",
-	"user_id": "bob",
-	"token_type": "bearer",
-	"expires_in": 86400
-}
