+++ System Tests/Networking/InstanceLocator.swift
@@ -13,11 +13,6 @@ struct InstanceLocator {
     // MARK: - Initializers
     
     init(_ instanceLocator: String) {
-        #if STUBBED
-        self.region = "STUBBED"
-        self.identifier = "STUBBED"
-        self.version = "STUBBED"
-        #else
         let components = instanceLocator.split(separator: InstanceLocator.separator)
         
         assert(components.count == 3, "Invalid format of the provided instance locator.")
@@ -25,7 +20,6 @@ struct InstanceLocator {
         self.region = String(components[1])
         self.identifier = String(components[2])
         self.version = String(components[0])
-        #endif
     }
     
 }
