+++ Unit Tests/Utilities/Bundle+Extensions.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+extension Bundle {
+    
+    // MARK: - Properties
+    
+    static let current = Bundle(for: BundleLocator.self)
+    
+}
+
+// MARK: - Bundle locator
+
+private class BundleLocator {}
